ack final fix i promise

+++ debugify.sh
@@ -4,7 +4,8 @@ echo Please select an option.
 echo 1. Install Debugify
 echo 2. Uninstall Debugify
 read option
-if [ $option == "1" ]; then
+if [ $option == "1" ]
+then
  echo "Installing Debugify Please Wait..."
  echo "Installing all dependencys"
  apt -y --allow-unauthenticated install wget
@@ -42,11 +43,12 @@ if [ $option == "1" ]; then
  echo
  echo "Successfully installed Debugify! Follow @B355P for more debug tweaks"
  ldrestart
-else
+elif [ $option == "2" ]
+then 
   echo "Uninstalling Debugify..."
   cd /
   rm -rf AppleInternal/
   echo "Change your SystemImageID back to finalize changes."
   ldrestart
-done
+fi
 end
